Don't set self.view.hidden = YES in popView.
In the case where there is only this view in the nav stack,
[self.navigationController popViewControllerAnimated:YES] doesn't
do anything, so setting self.view.hidden = YES causes the view to
go black in an ugly manner.

In the other case, setting self.view.hidden shouldn't be necessary,
because it's being popped off the screen (or whatever popViewBlock
decides to do).